+++ src/emu/nosv/setup.c
@@ -63,6 +63,7 @@ static struct ev_decl model_evlist[] = {
 	PAIR_E("VAl", "VAL", "nosv_mutex_lock()")
 	PAIR_E("VAt", "VAT", "nosv_mutex_trylock()")
 	PAIR_E("VAu", "VAU", "nosv_mutex_unlock()")
+	PAIR_E("VAb", "VAB", "nosv_barrier_wait()")
 
 	/* FIXME: VHA and VHa are not subsystems */
 	{ "VHa", "enters nosv_attach()" },
@@ -150,6 +151,7 @@ static const struct pcf_value_label nosv_ss_values[] = {
 	{ ST_API_MUTEX_LOCK,   "API: Mutex lock" },
 	{ ST_API_MUTEX_TRYLOCK,"API: Mutex trylock" },
 	{ ST_API_MUTEX_UNLOCK, "API: Mutex unlock" },
+	{ ST_API_BARRIER_WAIT, "API: Barrier wait" },
 	{ ST_WORKER,           "Thread: Worker" },
 	{ ST_DELEGATE,         "Thread: Delegate" },
 	{ EV_SCHED_SEND,       "EV Scheduler: Send task" },